refactor: dead code audit — delete 6 legacy ai_engine modules (-2,105 LOC)
Delete 6 files with zero callers or broken dependencies:
- classification/document_classifier.py (463 LOC) — legacy keyword
  classifier, 3 callers migrated to document_scanner.py
- ingestion/domain_ingest_orchestrator.py (825 LOC) — legacy
  orchestrator, imported from non-existent deal_intelligence_repo.py
- extraction/chunking.py (214 LOC) — thin wrapper around
  semantic_chunker.py, only caller was the deleted orchestrator
- ingestion/ingest_runner.py (157 LOC) — CLI wrapper, zero callers
- knowledge/knowledge_ingest_runner.py (524 LOC) — zero callers
- pdf/ic_cover.py (14 LOC) — stub marked for deletion

Migrate callers:
- document_scanner.py: absorb batch classification functions from
  deleted document_classifier.py (classify_documents,
  classify_registered_documents)
- extraction.py: inline reanalyze_deal (resolve_ai_mode +
  run_deal_ai_analysis) replacing orchestrator dependency
- pipeline_ingest_runner.py: remove Stage 4 orchestrator calls
  (already a runtime crash — deal_intelligence_repo.py never existed)

Fix broken imports discovered during audit:
- Create deal_intelligence_repo.py with register_deal_document() and
  update_deal_ai_output() — imported by registry_bridge.py and
  pipeline/persistence.py but module never existed
- Remove 3 dead worker functions from function_app.py referencing
  deleted operational modules (compliance, portfolio, reporting)
- Remove broken ai_review_analysis branch from memo_worker

Clean up:
- CLAUDE.md: remove stale references to deleted modules
- classification/__init__.py: update docstring
- document_scanner.py: consolidate hashlib import, remove dead guard,
  simplify _read_text_content query

337 tests pass. 5 import-linter contracts kept.
