--- backend/ai_engine/pdf/ic_cover.py
@@ -1,14 +1,0 @@
-# DELETE THIS FILE
-#
-# ic_cover.py was created during a refactoring session but is NOT needed.
-# All cover page logic lives in pdf_base.py (build_ic_cover_story).
-#
-# To delete:   del "D:\Projetos\Netz Private Credit OS\backend\ai_engine\pdf\ic_cover.py"
-# Or in PS:    Remove-Item "...\ic_cover.py"
-#
-# This stub raises an ImportError to prevent accidental use:
-
-raise ImportError(
-    "ic_cover.py is deprecated. "
-    "Import build_ic_cover_story from ai_engine.pdf.pdf_base instead.",
-)
